refactor: extract plugins
